Fix a few bugs
Due to the way Android animations work, a view will not be animated
unless it it on screen. Because of this the last group will never be
animated as it's child will never be on screen. To workaround this, we
give the child a initial height of 1.

Also on orientation change, the group info is wiped breaking the
collapse animation right after an orientation change. To fix this, we
set the initial value of the dummyHeight to -1. We can then
differentiate between a 0 and a -1 and use the full height if the value
of dummyHeight is -1.